Changed the box width and heights to be floats. This fixed the bug where there was extra space on the side of the screen if the numBoxesX was not a round number

+++ GeometryTunes/GridView.h
@@ -37,8 +37,8 @@
 - (void)changeNoteWithPitch:(unsigned)pitch octave:(unsigned)octave;
 - (void)changeNoteWithPitch:(unsigned)pitch octave:(unsigned)octave x:(unsigned) x y:(unsigned)y;
 
-- (int)getBoxWidth;
-- (int)getBoxHeight;
+- (float)getBoxWidth;
+- (float)getBoxHeight;
 
 - (void)drawGrid:(CGContextRef)context;
